backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pydantic==1.10.7
python-multipart==0.0.6
google-cloud-vision==3.4.0
python-dotenv==1.0.0 